+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Changed MSRV to 1.67.1.
 - Count physical cores when calculating database connection pool size.
+- Disallow duplicated poll options.
 
 ## [3.14.0] - 2025-01-14