+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Implemented poll publishing.
+- Added poll limits to `/api/v*/instance` response.
 
 ### Changed