+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Count physical cores when calculating database connection pool size.
 - Disallow duplicated poll options.
 - Preserve order of mentions.
+- Convert FEP-fb2a links to profile fields regardless of relation type.
 
 ## [3.14.0] - 2025-01-14