+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changed
 
 - Write warning to log when `content` of a vote is not null.
+- Changed MSRV to 1.68.2.
 
 ## Fixed